backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Database
sqlalchemy==2.0.35
asyncpg==0.29.0
alembic==1.13.2

# Redis and Celery
redis==5.1.1
celery[redis]==5.4.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.2

# Validation and settings
pydantic==2.9.2
pydantic-settings==2.5.2

# CORS
starlette==0.38.6

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
